Crate bolero_engine

Re-exports

pub use crate::test_failure::TestFailure;

Modules

panic

rng

shrink

test_failure

Structs

BorrowedGeneratorTest

BorrowedSliceTest

ByteSliceTestInput

ClonedGeneratorTest

ClonedSliceTest

DirectRngInput

Error

The Error type, a wrapper around a dynamic error type.

ForcedRngInput

SliceDebug

Enums

DriverMode

Byte exhaustion strategy for the driver

Traits

Driver

Trait for driving the generation of a value

Engine

Trait for defining an engine that executes a test

IntoTestResult

Trait that turns the test return value into a Result

Test

Trait for defining a test case

TestInput

TypeGenerator

Generate a value for a given type

ValueGenerator

Generate a value with a parameterized generator

Derive Macros

TypeGenerator

Derive the an implementation of TypeGenerator for the given type.